+++ test/test-webhook-handlers/webhook_test_handler.rb
@@ -1,12 +1,10 @@
-# This handler accepts webhooks from our integration tests. This webhook gets dispatched
-# if a banking provider test fails, indicating that the bank might be having an incident
-
 class WebhookTestHandler < Munster::BaseHandler
   def valid?(request)
     request.params.fetch(:isValid, false)
   end
 
   def process(webhook)
+    raise "Oops, failed" if webhook.request.params[:raiseDuringProcessing]
     filename = webhook.request.params.fetch(:outputToFilename)
     File.binwrite(filename, webhook.body)
   end